Avoid a build error for exe.wixproj between version 3.7 and 3.10
"exe.wixproj" execute a py launcher to merge NEWS files, but
a py launcher of python 3.11 with doubles quotes raises an error as below.

    > "py" -m pip install -U blurb
    Unable to create process using 'C:\hostedtoolcache\windows\Python\3.11.0\x64\py"  -m pip install -U blurb': The system cannot find the file specified.

This is why a build for exe.wixproj fails. To avoid a failure, use "py.exe"
instead of "py". Recently GitHub added python 3.11 into GitHub-Hosted runners,
so this failure appears.